requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0
groq
feedparser>=6.0.0
beautifulsoup4>=4.11.0
requests>=2.27.0